// src/types/cssClassEnums.ts
export enum BUTTON_PARTS {
  BUTTON = 'button',
  BUTTON_DISABLED = 'button-disabled',
  BUTTON_OUTLINED = 'button-outlined',
  BUTTON_OUTLINED_FOCUSED = 'button-outlined-focused',
  BUTTON_OUTLINED_DISABLED = 'button-outlined-disabled',
  BUTTON_START_ICON = 'button-start-icon',
  BUTTON_START_ICON_SMALL_WITH_PADDING = 'button-start-icon-small-with-padding',
  BUTTON_START_ICON_SMALL_WITHOUT_PADDING = 'button-start-icon-small-without-padding',
  BUTTON_START_ICON_MEDIUM_WITH_PADDING = 'button-start-icon-medium-with-padding',
  BUTTON_START_ICON_MEDIUM_WITHOUT_PADDING = 'button-start-icon-medium-without-padding',
  BUTTON_START_ICON_FAB = 'button-start-icon-fab',
  BUTTON_END_ICON = 'button-end-icon',
  BUTTON_START_ICON_RTL_MARGIN = 'button-start-icon-rtl-margin',
  BUTTON_TEXT = 'button-text',
  BUTTON_CONTAINED = 'button-contained',
  BUTTON_CONTAINED_INVERSE = 'button-contained-inverse',
  BUTTON_CONTAINED_DISABLED = 'button-contained-disabled',
  BUTTON_CONTAINED_DISABLED_INVERSE = 'button-contained-disabled-inverse',
  BUTTON_CONTAINED_HOVERED = 'button-contained-hovered',
  BUTTON_CONTAINED_FOCUSED = 'button-contained-focused',
  BUTTON_CONTAINED_FOCUSED_INVERSE = 'button-contained-focused-inverse',
  BUTTON_CONTAINED_HOVERED_FOCUSED = 'button-contained-hovered-and-focused',
  BUTTON_ENCHANTED_TEXT = 'button-enchanted-text',
  BUTTON_ENCHANTED_TEXT_INVERSE = 'button-enchanted-text-inverse',
  BUTTON_ENCHANTED_TEXT_DISABLED = 'button-enchanted-text-disabled',
  BUTTON_ENCHANTED_TEXT_DISABLED_INVERSE = 'button-enchanted-text-disabled-inverse',
  BUTTON_ENCHANTED_TEXT_HOVERED = 'button-enchanted-text-hovered',
  BUTTON_ENCHANTED_TEXT_FOCUSED = 'button-enchanted-text-focused',
  BUTTON_ENCHANTED_TEXT_FOCUSED_INVERSE = 'button-enchanted-text-focused-inverse',
  BUTTON_ENCHANTED_TEXT_HOVERED_FOCUSED = 'button-enchanted-text-hovered-and-focused',
  BUTTON_ENCHANTED_OUTLINED = 'button-enchanted-outlined',
  BUTTON_ENCHANTED_OUTLINED_FOCUSED = 'button-enchanted-outlined-focused',
  BUTTON_ENCHANTED_OUTLINED_DISABLED = 'button-enchanted-outlined-disabled',
  BUTTON_ENCHANTED_OUTLINED_HOVERED = 'button-enchanted-outlined-hovered',
  BUTTON_ENCHANTED_OUTLINED_HOVERED_FOCUSED = 'button-enchanted-outlined-hovered-and-focused',
}
// ...
export enum AVATAR_PARTS {
  AVATAR_DIV = 'avatar-div',
  AVATAR_DIV_CIRCULAR = 'avatar-div-circular',
  AVATAR_SPAN_CIRCULAR = 'avatar-span-circular',
  AVATAR_SPAN_ROUNDED = 'avatar-span-rounded',
  AVATAR_ICON_CIRCULAR = 'avatar-icon-circular',
  AVATAR_ICON_ROUNDED = 'avatar-icon-rounded',
  AVATAR_ICON_TEMPLATE_CIRCULAR = 'avatar-icon-template-circular',
  AVATAR_ICON_TEMPLATE_ROUNDED = 'avatar-icon-template-rounded',
  AVATAR_IMAGE_CIRCULAR = 'avatar-image-circular',
  AVATAR_IMAGE_ROUNDED = 'avatar-image-rounded',
}

export enum AVATAR_COLOR {
  AVATAR_DEFAULT_COLOR = 'avatar-default-color',
  AVATAR_RED = 'avatar-red',
  AVATAR_ORANGE = 'avatar-orange',
  AVATAR_YELLOW = 'avatar-yellow',
  AVATAR_LIME = 'avatar-lime',
  AVATAR_GREEN = 'avatar-green',
  AVATAR_TEAL = 'avatar-teal',
  AVATAR_BLUE = 'avatar-blue',
  AVATAR_INDIGO = 'avatar-indigo',
  AVATAR_PURPLE = 'avatar-purple',
  AVATAR_PINK = 'avatar-pink',
}
// ...
export enum LIST_PARTS {
  UNORDERED_LIST = 'unordered-list',
}

export enum LIST_ITEM_PARTS {
  LIST_ITEM = 'list-item',
  LIST_ITEM_SELECTED = 'list-item-selected',
  MENU_ITEM = 'menu-item',
}
// ...
export enum INPUT_SELECT_PARTS {
  DIV = 'div',
  DIV_LABEL = 'div-label',
  LABEL = 'label',
  LABEL_DISABLED = 'label-disabled',
  REMOVE_LABEL = 'remove-label',
  REMOVE_LABEL_DISABLED = 'remove-label-disabled',
}
// ...
export enum TOOLTIP_PARTS {
  TOOLTIP_ROOT = 'tooltip-root',
  TOOLTIP_TEXT = 'tooltip-text',
  TOOLTIP_TEXT_HIDDEN = 'tooltip-text-hidden',
  TOOLTIP_TARGET = 'tooltip-target',
}

export enum TOOLTIP_VARIANT {
  TOOLTIP_SMALL = 'tooltip-small',
  TOOLTIP_MEDIUM = 'tooltip-medium',
}

export enum TOOLTIP_TYPE {
  TOOLTIP_SINGLE_LINE = 'tooltip-single-line',
  TOOLTIP_MULTI_LINE = 'tooltip-multi-line',
}

export enum TOOLTIP_PLACEMENT {
  TOOLTIP_BOTTOM = 'tooltip-bottom',
  TOOLTIP_BOTTOM_START = 'tooltip-bottom-start',
  TOOLTIP_BOTTOM_END = 'tooltip-bottom-end',
  TOOLTIP_TOP = 'tooltip-top',
  TOOLTIP_TOP_START = 'tooltip-top-start',
  TOOLTIP_TOP_END = 'tooltip-top-end',
  TOOLTIP_RIGHT = 'tooltip-right',
  TOOLTIP_RIGHT_START = 'tooltip-right-start',
  TOOLTIP_RIGHT_END = 'tooltip-right-end',
  TOOLTIP_LEFT = 'tooltip-left',
  TOOLTIP_LEFT_START = 'tooltip-left-start',
  TOOLTIP_LEFT_END = 'tooltip-left-end',
}